requirements.txt
discord.py
pystyle
colorama
